When a file contains several maps, look for a default map
Rather than always taking the first one in the file. This is exactly the
intended use of the 'default' flag.

Note that pretty much the same code is duplicated in xkbcomp.c when
compiling a single file from the command line, but there it is
implemented correctly (look for XkbLC_Default).

https://bugs.freedesktop.org/show_bug.cgi?id=69950

Signed-off-by: Ran Benita <ran234@gmail.com>
Tested-by: Benno Schulenberg <bensberg@justemail.net>
Acked-by: Sergey Udaltsov <sergey.udaltsov@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>