Downgrade the warning for missing symbols to info
In the interest of maintainability, it's easier to include as many keycodes as
possible and then have the symbols mapping specific to the layout. This is
particularly true for evdev where the kernel takes care of device-specifics
and every keyboard has the same set of keycodes anyway.

So let's downgrade this from a warning to a mere info, virtually every
keyboard right now triggers this warning for a number of keys.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>